refactor(py): moved ros_bridge unit tests to bridges/ros and updated docstring from ros_bridge -> bridges.ros
