+++ mosaico-sdk-py/pyproject.toml
@@ -68,7 +68,7 @@ requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-"mosaicolabs.ros_injector" = "mosaicolabs.ros_bridge.injector:ros_injector"
+"mosaicolabs.ros_injector" = "mosaicolabs.bridges.ros.injector:ros_injector"
 "mosaicolabs.examples" = "mosaicolabs.examples.cli:run_example_cli"
 "mosaicolabs.testing" = "testing.testing:mosaico_testing"
 "mosaico" = "mosaicolabs_cli.main:app"
