+++ mosaico-sdk-py/src/testing/unit/bridges/ros/test_registration.py
