+++ mosaico-sdk-py/src/testing/unit/bridges/ros/test_registry_scoping.py
