+++ mosaico-sdk-py/src/testing/unit/bridges/ros/test_ros_schema_metadata.py
