+++ mosaico-sdk-py/src/testing/unit/bridges/ros/test_schema_metadata_from_injector.py
