+++ mosaico-sdk-py/src/testing/unit/bridges/ros/test_typestore_registration.py
